link fixes, state management fix, final
