+++ src/components/Reddit/RedditContent.css
@@ -21,8 +21,4 @@
     filter: drop-shadow(0 -12px 12px);
     padding: 1em 2.5em 0em 2.5em;
 }
-.divider{
-  margin-top: 35px;
-  padding-bottom: 20px;
-  border-top: 2px solid black;
-}
+
